test: clean up and no parsing of errInfo
